+++ lib/pages/home/home_page.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import 'api.dart';
-import 'cadastrar_tarefa.dart';
-import 'tarLista.dart';
+import '../../models/tar_lista.dart';
+import '../../services/api.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -12,16 +11,13 @@ class HomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<HomePage> with TickerProviderStateMixin {
-  @override
   List<TarLista> tarefas = [];
   bool _isPlay = false;
   late AnimationController _controller;
-  
 
   @override
   void initState() {
-     _controller =
-        AnimationController(duration: const Duration(seconds: 1), vsync: this);
+    _controller = AnimationController(duration: const Duration(seconds: 1), vsync: this);
     super.initState();
     API.listaTarefas().then((items) {
       setState(() {
@@ -31,7 +27,7 @@ class _MyHomePageState extends State<HomePage> with TickerProviderStateMixin {
   }
 
   //dispose para o botão de play
-   @override
+  @override
   void dispose() {
     _controller.dispose();
     super.dispose();
@@ -43,7 +39,7 @@ class _MyHomePageState extends State<HomePage> with TickerProviderStateMixin {
       drawer: Drawer(
         child: Column(
           children: [
-            Spacer(),
+            const Spacer(),
             ElevatedButton(
               onPressed: () async {
                 bool logOut = await logout();
@@ -53,7 +49,7 @@ class _MyHomePageState extends State<HomePage> with TickerProviderStateMixin {
               },
               child: const Text('Deslogar'),
             ),
-            SizedBox(height: 40),
+            const SizedBox(height: 40),
           ],
         ),
       ),
@@ -73,42 +69,38 @@ class _MyHomePageState extends State<HomePage> with TickerProviderStateMixin {
           ),
           Expanded(
             child: ListView(
-              
               children: tarefas.map(
                 (tarefa) {
                   return Card(
                     child: ListTile(
-                      
                       leading: GestureDetector(
                         onTap: () {
-                          
-                          
                           if (_isPlay == false) {
-                _controller.forward();
-                _isPlay = true;
-              } else {
-                _controller.reverse();
-                _isPlay = false;
-              }
+                            _controller.forward();
+                            _isPlay = true;
+                          } else {
+                            _controller.reverse();
+                            _isPlay = false;
+                          }
                         },
                         child: AnimatedIcon(
                           icon: AnimatedIcons.play_pause,
                           progress: _controller,
-                          color: Colors.blue,),
+                          color: Colors.blue,
+                        ),
                       ),
                       title: Text(tarefa.nome),
                       trailing: Row(
                         mainAxisSize: MainAxisSize.min,
                         children: [
-                          IconButton(onPressed: () {
-                            print("editar tarefa ");
-                          }, 
-                          icon: Icon(Icons.edit)),
-                          IconButton(onPressed: () {}, 
-                          icon: Icon(Icons.delete)),
+                          IconButton(
+                              onPressed: () {
+                                print("editar tarefa ");
+                              },
+                              icon: const Icon(Icons.edit)),
+                          IconButton(onPressed: () {}, icon: const Icon(Icons.delete)),
                         ],
                       ),
-                      
                     ),
                   );
                 },
